Full support for evidences of instance measures
Includes full support for returning evidences of instance measures together with multi-instance measures. The current implementation removes the need to specify it in the measure and allows a flexible configuration by means of ComputerConfig. Specifically, evidences can be computed for all measure definitions or just for those measure definitions that have an id. It also works together with overrides.

Evidences can be flattened, i.e. all included in the last measure, (default behavior) or they can be placed in the measure that directly used them for their computation.

Also, some refactoring was made to remove Overrides since it was redundant with the new ComputerConfig.